Move ensure_file_obj() to converters module
